// types.ts
export interface Filters {
  discoveryFacilities: string[];
  detectionMethod: string;
  planetTypes: string[];
  hostName: string;
  /** When in IA mode, filter rows generated by AI */
  aiGeneratedOnly?: boolean;
}

export interface Exoplanet {
  // Identification
  kepoi_name: string;
  kepler_name: string | null;
  kepid: number;
  
  // Disposition et statut
  koi_disposition: string;
  koi_pdisposition: string | null;
  koi_score: number | null;
  
  // Paramètres orbitaux
  koi_period: number | null;
  koi_period_err1: number | null;
  koi_period_err2: number | null;
  koi_sma: number | null;
  koi_sma_err1: number | null;
  koi_sma_err2: number | null;
  koi_eccen: number | null;
  koi_eccen_err1: number | null;
  koi_eccen_err2: number | null;
  koi_incl: number | null;
  koi_incl_err1: number | null;
  koi_incl_err2: number | null;
  
  // Paramètres physiques de la planète
  koi_prad: number | null;
  koi_prad_err1: number | null;
  koi_prad_err2: number | null;
  koi_teq: number | null;
  koi_teq_err1: number | null;
  koi_teq_err2: number | null;
  koi_insol: number | null;
  koi_insol_err1: number | null;
  koi_insol_err2: number | null;
  
  // Paramètres de transit
  koi_depth: number | null;
  koi_depth_err1: number | null;
  koi_depth_err2: number | null;
  koi_duration: number | null;
  koi_duration_err1: number | null;
  koi_duration_err2: number | null;
  koi_ror: number | null;
  koi_ror_err1: number | null;
  koi_ror_err2: number | null;
  koi_impact: number | null;
  koi_impact_err1: number | null;
  koi_impact_err2: number | null;
  
  // Paramètres stellaires
  koi_steff: number | null;
  koi_steff_err1: number | null;
  koi_steff_err2: number | null;
  koi_slogg: number | null;
  koi_slogg_err1: number | null;
  koi_slogg_err2: number | null;
  koi_smet: number | null;
  koi_smet_err1: number | null;
  koi_smet_err2: number | null;
  koi_srad: number | null;
  koi_srad_err1: number | null;
  koi_srad_err2: number | null;
  koi_smass: number | null;
  koi_smass_err1: number | null;
  koi_smass_err2: number | null;
  
  // Coordonnées
  ra: number | null;
  dec: number | null;
  ra_str: string | null;
  dec_str: string | null;
  
  // Magnitudes
  koi_kepmag: number | null;
  koi_gmag: number | null;
  koi_rmag: number | null;
  koi_imag: number | null;
  koi_jmag: number | null;
  koi_hmag: number | null;
  koi_kmag: number | null;
  
  // Informations supplémentaires
  koi_count: number | null;
  koi_num_transits: number | null;
  koi_quarters: string | null;
  koi_time0: number | null;
  koi_time0_err1: number | null;
  koi_time0_err2: number | null;
  koi_comment: string | null;
  koi_vet_date: string | null;
}

export enum PlanetType {
    ROCKY = 'Rocky',
    SUPER_EARTH = 'Super-Earth / Mini-Neptune',
    GAS_GIANT = 'Gas Giant',
}